Update 20/08/2021 By CJ - Code Rework
Added an Interface class for interaction of Simulation between classes,
as well as an abstract Actor class for all interactable objects within
the simulation to inherit from.

PackingStation was reworked as Order already contains details such as
Number of Ticks to pack, etc.

ChargingPod and Robot partially interactable with one another, and a
State added to Robots to caculate how much power is being used depending
on the actions taken.

Position added to help calculate the x and y coordinates of the
interactable (Actor) objects.