20/08/2021 - Code Rework
Added an Interface Entity to connect classes to the Simulator in the
future, as well as an Absract Actor class with the basic information
(UID etc) that every interactable object within the simulation requires,
and can inherit from.

Changes were made to the PackingStation class due to already having
NumberOfTicks and Status as part of the individual Orders. The Enum
State refers to the Robot and how much power it uses for actions while
Position tells the simulation where each individual object is within the
grid (still needs warehouse intergration).