20/08/2021 - Basic Code Upload
Entity Interface added for future connection to the Simulator and
Warehouse. Abstract Actor entity created and used to inherit for all
interactable objects (robot etc). Basic code added to ChargingPod, and
more thorough code added to Robot including an Enum State to show how
much power would be used for activities (actions not yet implemented).

PackingStation changed due to having NumberOfTicks and Status within the
Order class. Position added to be connected with the Warehouse, giving
the Simulation the position of each object active on the grid.